| more refactoring | removing needless function | dividing class to different folder | adding more doc comment |
