+++ Assets/Unity-Open-Stories/Scripts/Singleton/Comic.cs
@@ -85,5 +85,20 @@ namespace as3mbus.Story
                 return names;
             }
         }
+        // return list of folder containing file with image extension inside an asset bundle
+        public static string[] getComics(AssetBundle comicBundle)
+        {
+            List<string> comBunCon = new List<string>();
+            foreach (string asset in comicBundle.GetAllAssetNames())
+            {
+                if ((asset.Contains(".png") || asset.Contains(".jpg") || asset.Contains(".jpeg")) && !comBunCon.Contains(Path.GetDirectoryName(asset)))
+                {
+                    comBunCon.Add(Path.GetDirectoryName(asset));
+                    // Debug.Log(Path.GetDirectoryName(asset));
+                }
+            }
+            comicBundle.Unload(false);
+            return comBunCon.ToArray();
+        }
     }
 }
