+++ Assets/Unity-Open-Stories/Scripts/Singleton/JsonKey.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: e7f619e40fa3e49b5860834dff893397
+timeCreated: 1512741815
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
