+++ Assets/Unity-Open-Stories/Scripts/Singleton/StoryJsonKey.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 38bddbb39e1b54235858e029f174ebfb
+timeCreated: 1512741814
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
